feat: Set up Fastapi server and migrates all core pipeline methods to src/services/
