+++ src/services/__init__.py
@@ -1,0 +1,3 @@
+'''
+Module containing functions to handle the RAG Pipeline
+'''
